add function to import data from JSON & CSV into db
The CLI can now be used to import a full data set in one go
(though it should not try to create the database every time...
or the creation and insertion should be split into separate
subcommands).

The DatasetX test data have been updated to reflect the expected format.

NB! The code is currently broken!
The data in the CSV need to be linked to the correct sample in the
Samples table, but the foreign key points to an autoincrementing int.
One could perhaps create an ID:SampleName mapping after inserting the
sample metadata, then modify the CSV data before insertion.