fix function to import data from JSON & CSV into db
The data in the CSV is now linked to the correct samples in the
Samples table. The foreign key 'SampleID' in the metrics table
successfully points to the SampleID used in the Samples table,
which is an autoincrementing int.

To achieve the link between SampleIDs of these two tables (and
indeed in future cases), the insert_into_db fn was modifed to
return a list of the last row ids inserted.