add error handling and rollback of changes made to the database.
create_db fn is wrapped in a try-except block to allow the program
to continue if the database already exists. This is currently
necessary since the creation of the db and inserting data have
not yet been decoupled.

add_data_to_db is wrapped in a try-except block to allow for
changes to the database to be rolled back should an error occur
at any stage in the addition of data to tables. For this to be
possible, the db.transaction method has been introduced to
keep track of changes made to the db. This can be used by
calling either t.commit() or t.rollback() where applicable.

The database connection object has also been moved to the top
level to allow for greater control over database transactions.

Note that tests have not yet been written for these changes.